requirements.txt
# VAPT Pipeline Dependencies

# Core dependencies
pyyaml>=6.0
requests>=2.28.0
jinja2>=3.1.0

# Faraday CLI (uncomment to use in CI/CD)
# faraday-cli>=2.4.0

# Optional: For NVD API enrichment (Bonus Challenge)
# nist-data-mirror>=0.0.33

# Optional: For dashboard (Bonus Challenge B)
# flask>=2.3.0
# flask-cors>=4.0.0
# pandas>=1.5.0
# plotly>=5.14.0

# Optional: For false positive filtering (Bonus Challenge C)
# scikit-learn>=1.2.0
# sqlite3 (built-in)

# Development dependencies
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=4.0.0
# mypy>=0.990
